+++ res_partner_fiscal_position_by_country/__manifest__.py
@@ -10,7 +10,9 @@
     "depends": [
         "account",
     ],
-    "data": [],
+    "data": [
+        "views/views.xml"
+    ],
     "installable": True,
     "auto_install": True,
 }
