+++ res_partner_fiscal_position_by_country/models/__init__.py
@@ -1,1 +1,3 @@
+from . import account_fiscal_position
+from . import res_country
 from . import res_partner
