+++ res_partner_fiscal_position_by_country/models/res_country.py
@@ -1,0 +1,9 @@
+
+from odoo import api, fields, models
+
+
+class ResCountry(models.Model):
+    _inherit = 'res.country'
+
+    default_incoterm_id = fields.Many2one(
+        "account.incoterms", "Default Incoterm")
